refactor: 무한스크롤 위치 변경, store post랑stack통합
