+++ src/components/PostList.js
@@ -10,6 +10,7 @@ import dummyPost from "../dummyPost/dummyPost.json";
 import { useDispatch, useSelector } from "react-redux";
 import { loadMorePostListAsync, loadPostListAsync } from "../store/post/post";
 import Loading from "../pages/Loading";
+import { useIntersectionObserver } from "../hooks/useIntersectionObserver";
 // // Main page안의 카드 슬롯 형태 리스트
 
 const BigTable = styled.div`
@@ -57,47 +58,18 @@ function a11yProps(index) {
 
 export default function BasicTabs() {
   const dispatch = useDispatch();
-  const [observationTarget, setObservationTarget] = useState(null);
   const [value, setValue] = useState(0);
-
-  const firstLoading = useSelector(state => state.post.FirstPostListDone);
   const isLoading = useSelector(state => state.post.MorePostListLoading);
   const stackList = useSelector(state => state.post.stackList);
   const postListData = useSelector(state => state.post.postListData);
-  let postListShow = useSelector(state => state.post.postListShow);
+  const postListShow = useSelector(state => state.post.postListShow);
   const handleChange = (event, newValue) => {
     setValue(newValue);
   };
   useEffect(() => {
     dispatch(loadPostListAsync());
-    console.log("초기불러오기 완료");
   }, [dispatch]);
-  // -----------------------------------------------------------------------
-  const observer = useRef(
-    new IntersectionObserver(
-      ([entry]) => {
-        if (!entry.isIntersecting) return;
-        //함수자리
-        console.log("발생");
-        dispatch(loadMorePostListAsync());
-      },
-      { threshold: 1 }
-    )
-  );
 
-  useEffect(() => {
-    const currentTarget = observationTarget;
-    const currentObserver = observer.current;
-    if (currentTarget) {
-      currentObserver.observe(currentTarget);
-    }
-    return () => {
-      if (currentTarget) {
-        currentObserver.unobserve(currentTarget);
-      }
-    };
-  }, [observationTarget]);
-  // -----------------------------------------------------------------------
   return (
     <Box sx={{ width: "100%" }}>
       <Box sx={{ borderBottom: 1, borderColor: "divider" }}>
@@ -115,42 +87,60 @@ export default function BasicTabs() {
       <TabPanel value={value} index={0}>
         <BigTable>
           <Table>
-            {firstLoading &&
-              postListData.map(data => {
-                if (stackList.length === 0) {
+            {postListData.map(data => {
+              if (stackList.length === 0) {
+                return (
+                  <PostPreview
+                    title={data.title}
+                    start_date={data.start_date}
+                    designated_stacks={data.designated_stacks}
+                    poster_nickname={data.poster_nickname}
+                    post_id={data.post_id}
+                  />
+                );
+              } else {
+                const stackChecked = data.designated_stacks.filter(x =>
+                  stackList.includes(x)
+                );
+                if (!(stackChecked.length === 0)) {
                   return (
                     <PostPreview
                       title={data.title}
                       start_date={data.start_date}
                       designated_stacks={data.designated_stacks}
                       poster_nickname={data.poster_nickname}
-                      post_id={data.post_id}
                     />
                   );
-                } else {
-                  const stackChecked = data.designated_stacks.filter(x =>
-                    stackList.includes(x)
-                  );
-                  if (!(stackChecked.length === 0)) {
-                    return (
-                      <PostPreview
-                        title={data.title}
-                        start_date={data.start_date}
-                        designated_stacks={data.designated_stacks}
-                        poster_nickname={data.poster_nickname}
-                      />
-                    );
-                  }
                 }
-              })}
+              }
+            })}
             {isLoading && <Loading />}
-            {!isLoading && <div ref={setObservationTarget}></div>}
           </Table>
         </BigTable>
       </TabPanel>
+
       <TabPanel value={value} index={1}>
-        스터디
+        <BigTable>
+          <Table>
+            {postListShow[0] ? (
+              postListShow.map(data => {
+                return (
+                  <PostPreview
+                    title={data.title}
+                    start_date={data.start_date}
+                    designated_stacks={data.designated_stacks}
+                    poster_nickname={data.poster_nickname}
+                    post_id={data.post_id}
+                  />
+                );
+              })
+            ) : (
+              <div>빈페이지</div>
+            )}
+          </Table>
+        </BigTable>
       </TabPanel>
+
       <TabPanel value={value} index={2}>
         프로젝트
       </TabPanel>
