+++ src/components/StackButton/StackButton.js
@@ -5,6 +5,7 @@ import { addStack, deleteStack } from "../../store/post/post";
 const StackButton = ({ stack }) => {
   const dispatch = useDispatch();
   const stackLists = useSelector(state => state.post.stackList);
+  const postListShow = useSelector(state => state.post.postListShow);
   return (
     <ToggleButton
       sx={{ borderRadius: 10, marginRight: 2 }}
@@ -13,6 +14,7 @@ const StackButton = ({ stack }) => {
       onChange={() => {
         if (!stackLists.includes(`${stack}`)) {
           dispatch(addStack({ stack: stack }));
+          console.log(postListShow);
         }
         if (stackLists.includes(`${stack}`)) {
           dispatch(deleteStack({ stack: stack }));
