+++ src/hooks/useIntersectionObserver.js
@@ -1,0 +1,32 @@
+import { useEffect, useRef, useState } from "react";
+import { useDispatch } from "react-redux";
+import { loadMorePostListAsync } from "../store/post/post";
+
+export const useIntersectionObserver = () => {
+  const dispatch = useDispatch();
+  const [observationTarget, setObservationTarget] = useState(null);
+  const observer = useRef(
+    new IntersectionObserver(
+      ([entry]) => {
+        if (!entry.isIntersecting) return;
+        dispatch(loadMorePostListAsync());
+      },
+      { threshold: 1 }
+    )
+  );
+
+  useEffect(() => {
+    const currentTarget = observationTarget;
+    const currentObserver = observer.current;
+    if (currentTarget) {
+      currentObserver.observe(currentTarget);
+    }
+    return () => {
+      if (currentTarget) {
+        currentObserver.unobserve(currentTarget);
+      }
+    };
+  }, [observationTarget]);
+
+  return setObservationTarget;
+};
